fix waPDF for global object
