+++ index.html
@@ -10,7 +10,6 @@
   <!-- TODO: 之後換成 mqtt.js -->
   <script src="https://cdnjs.cloudflare.com/ajax/libs/paho-mqtt/1.1.0/paho-mqtt.min.js"></script>
   <script src="https://mozilla.github.io/pdf.js/build/pdf.js"></script>
-  <script src="./waPDF.js"></script>
 
 </head>
