+++ src/waPDF.d.ts
@@ -1,0 +1,2 @@
+declare const waPDF: any; // or your specific type
+export default waPDF;
